+++ src/routes/notes.js
@@ -4,16 +4,16 @@ const router = express.Router();
 // Mongoose.Model de Nota para la base de datos
 // Crea una clase que se puede instanciar y utilizar.
 const Note = require('../models/Note');
-
+const { isAuthenticated } = require('../helpers/auth'); // Comprobará si está logueado o no
 
 // Añade nueva nota
-router.get('/notes/add', (req, res) => {
+router.get('/notes/add', isAuthenticated, (req, res) => {
   res.render('notes/new-note');
 });
 
 
 // Recibiendo datos de formulario 'New-Note'
-router.post('/notes/new-note', async (req, res) => {
+router.post('/notes/new-note', isAuthenticated, async (req, res) => {
   const { title, description } = req.body;
   const errors = [];
   if (!title) {
@@ -38,12 +38,12 @@ router.post('/notes/new-note', async (req, res) => {
 
 
 // Editando notas
-router.get('/notes/edit/:id', async (req, res) => {
+router.get('/notes/edit/:id', isAuthenticated, async (req, res) => {
   const note = await Note.findById(req.params.id);
   res.render('notes/edit-note', { note });
 });
 // ... que envía un PUT recibido por la siguiente
-router.put('/notes/edit-note/:id', async (req, res) => {
+router.put('/notes/edit-note/:id', isAuthenticated, async (req, res) => {
   const { title, description } = req.body;
   await Note.findByIdAndUpdate(req.params.id, { title, description });
   req.flash('success_msg', 'Nota editada...');
@@ -52,7 +52,7 @@ router.put('/notes/edit-note/:id', async (req, res) => {
 
 
 // Borrar nota
-router.delete('/notes/delete/:id', async (req, res) => {
+router.delete('/notes/delete/:id', isAuthenticated, async (req, res) => {
   await Note.findByIdAndDelete(req.params.id);
   req.flash('success_msg', 'Nota eliminada...');
   res.redirect('/notes');
@@ -60,7 +60,7 @@ router.delete('/notes/delete/:id', async (req, res) => {
 
 
 // Muestra todas las notas
-router.get('/notes', async (req, res) => {
+router.get('/notes', isAuthenticated, async (req, res) => {
   const notes = await Note.find().sort({ date: 'desc' });
   res.render('notes/all-notes', { notes });
 });