+++ src/routes/users.js
@@ -4,8 +4,8 @@
 const express = require('express');
 const router = express.Router();
 
-const User = require('../models/User');
 const passport = require('passport');
+const User = require('../models/User');
 
 
 // Formulario Ingresando usuario
@@ -21,7 +21,7 @@ router.get('/users/signin', (req, res) => {
 router.post('/users/signin', passport.authenticate('local', {
   successRedirect: '/notes',
   failureRedirect: '/users/signin',
-  failureFlash: true
+  failureFlash: true,
 }));
 
 
@@ -59,6 +59,13 @@ router.post('/users/signup', async (req, res) => {
     req.flash('success_msg', 'Estas registrado...');
     res.redirect('/users/signin');
   }
+});
+
+
+// Logout. Gestión de fin de sesión
+router.get('/users/logout', (req, res) => {
+  req.logout();
+  res.redirect('/');
 });